Move FakeFlag and FakeS3Event over to tests/classes_mock.py
